api/requirements.txt
flask==2.3.3
mysql-connector-python==8.4.0
cryptography==42.0.8
python-dotenv==1.0.1
numpy==1.26.4
